move check of redacted/omitted fields to config class
